admin-backend/script/requirements.txt
bcrypt>=4.0.0
psycopg2-binary>=2.9.0
pyyaml>=6.0
minio>=7.2.0
